requirements.txt
python-telegram-bot>=20.6

python-dotenv
supabase
httpx
aiohttp
pytz>=2023.3

